+++ packages/proteus/src/main/util/index.ts
@@ -17,9 +17,10 @@
  *
  */
 
-export {NewTeamInvitation, TeamInvitation, TeamInvitationAPI, TeamInvitationChunk} from './invitation/';
-export {MemberAPI, MemberData, Members, PermissionsData} from './member/';
-export {NewTeamData, TeamAPI, TeamChunkData, TeamData, TeamInfo} from './team/';
-export {PaymentAPI, PaymentData} from './payment/';
-export {ServiceAPI, Service, ServiceWhitelistData} from './service/';
-export {TeamError, InviteEmailInUseError, InvalidInvitationCodeError, ServiceNotFoundError} from './TeamError';
+import ArrayUtil from './ArrayUtil';
+import ClassUtil from './ClassUtil';
+import KeyDerivationUtil from './KeyDerivationUtil';
+import MemoryUtil from './MemoryUtil';
+import RandomUtil from './RandomUtil';
+
+export {ArrayUtil, ClassUtil, KeyDerivationUtil, MemoryUtil, RandomUtil};